// web/src/domain/types/sessions.ts
export enum SessionActionTypes { 
    DISPATCH_SIGN_IN_ACTION = 'DISPATCH_SIGN_IN_ACTION',
    SIGN_IN_SUCCESS_ACTION = 'SIGN_IN_SUCCESS_ACTION',
    SIGN_IN_FAILURE_ACTION = 'SIGN_IN_FAILURE_ACTION',

    DISPATCH_SIGN_OUT_ACTION = 'DISPATCH_SIGN_OUT_ACTION',
    SIGN_OUT_SUCCESS_ACTION = 'SIGN_OUT_SUCCESS_ACTION',

    DISPATCH_VALIDATE_SESSION_ACTION = 'DISPATCH_VALIDATE_SESSION_ACTION',
    VALIDATE_SESSION_SUCCESS_ACTION = 'VALIDATE_SESSION_SUCCESS_ACTION',
    VALIDATE_SESSION_FAILURE_ACTION = 'VALIDATE_SESSION_FAILURE_ACTION',

    DISPATCH_VALIDATE_SESSION_PERMISSIONS_ACTION = 'DISPATCH_VALIDATE_SESSION_PERMISSIONS_ACTION',
    SET_VALIDATED_SESSION_PERMISSIONS_ACTION = 'SET_VALIDATED_SESSION_PERMISSIONS_ACTION'
}


export enum UserSignInStatus {
    IDLE = 'IDLE',
    SUCCESS = 'SUCCESS',
    PENDING = 'PENDING',
    FAILURE = 'FAILURE'
  }